# api-spec.yaml
openapi: 3.0.3
info:
  title: VetCare API
  description: API for veterinary clinic management system
  version: 1.0.0
  contact:
    name: VetCare API Support
servers:
  - url: https://api.vetcare.example.com
    description: Production server
  - url: http://localhost:5000
    description: Development server
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login or /api/auth/register
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Error message (alternative field name)

    # Auth schemas
    RegisterDto:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "owner@example.com"
        password:
          type: string
          format: password
          example: "SecurePassword123!"
        fullName:
          type: string
          nullable: true
          example: "John Doe"
    LoginDto:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "owner@example.com"
        password:
          type: string
          format: password
          example: "SecurePassword123!"
    RefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expiresAtUtc:
          type: string
          format: date-time
          example: "2024-12-31T23:59:59Z"
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    CreateClinicStaffDto:
      type: object
      required:
        - email
        - password
        - clinicId
      properties:
        email:
          type: string
          format: email
          example: "staff@clinic.com"
        password:
          type: string
          format: password
          example: "SecurePassword123!"
        fullName:
          type: string
          nullable: true
          example: "Jane Smith"
        clinicId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
    ClinicStaffResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
        clinicId:
          type: string
          format: uuid
          nullable: true

    # Clinic schemas
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user (owner)
      description: Public endpoint for owner self-registration
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
            example:
              email: "owner@example.com"
              password: "SecurePassword123!"
              fullName: "John Doe"
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
              example:
                accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3OC0xMjM0LTU2NzgtODkwMS0xMjM0NTY3ODkwMTIiLCJyb2xlIjoiVXNlciIsImV4cCI6MTczNDU2Nzg5MH0.example"
                expiresAtUtc: "2024-12-20T12:00:00Z"
                refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh_token_example"
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Email already registered"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      description: Public endpoint for user authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
            example:
              email: "owner@example.com"
              password: "SecurePassword123!"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
              example:
                accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3OC0xMjM0LTU2NzgtODkwMS0xMjM0NTY3ODkwMTIiLCJyb2xlIjoiVXNlciIsImV4cCI6MTczNDU2Nzg5MH0.example"
                expiresAtUtc: "2024-12-20T12:00:00Z"
                refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh_token_example"
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Invalid credentials"
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Public endpoint to refresh access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
            example:
              refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh_token_example"
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
              example:
                accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.new_access_token"
                expiresAtUtc: "2024-12-20T12:00:00Z"
                refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.new_refresh_token"
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Invalid or expired refresh token"
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout and revoke refresh token
      description: Revokes the current refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
            example:
              refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh_token_example"
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out successfully"
  /api/auth/admin-check:
    get:
      tags: [Auth]
      summary: Check admin access
      description: Admin-only endpoint to verify admin role
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Admin access confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  role:
                    type: string
                    example: "Admin"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - not an admin
  /api/auth/create-clinic-staff:
    post:
      tags: [Auth]
      summary: Create clinic staff user (Admin only)
      description: Admin-only endpoint to create a clinic staff user for a specific clinic
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClinicStaffDto'
            example:
              email: "staff@clinic.com"
              password: "SecurePassword123!"
              fullName: "Jane Smith"
              clinicId: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Clinic staff created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClinicStaffResponse'
              example:
                id: "123e4567-e89b-12d3-a456-426614174004"
                fullName: "Jane Smith"
                email: "staff@clinic.com"
                role: "ClinicStaff"
                clinicId: "123e4567-e89b-12d3-a456-426614174000"
        '404':
          description: Clinic not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Clinic not found"
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Email already registered"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - not an admin

  # Clinic endpoints
...